Use correct types
